requirements.txt
streamlit
pillow
exifread
streamlit-folium
folium
geopy
requests
